requirements.txt
python-telegram-bot==11.1.0
googletrans==2.4.0
